Revert "Merge SessionFactory into SessionRepository"
